Promjenjeni intovi u float #7

+++ Software/app/src/main/java/hr/foi/rampu/fridgium/entities/NamirnicaPrikaz.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 data class NamirnicaPrikaz(
     val id: Int,
     val naziv: String,
-    val kolicina_hladnjak: Int,
+    val kolicina_hladnjak: Float,
      @SerializedName("mjerna_jedinica_id") val mjernaJedinica: MjernaJedinica,
-    val kolicina_kupovina: Int,
-    var kolicina: Int
+    val kolicina_kupovina: Float,
+    var kolicina: Float
 )
